requirements.txt
# API通信
requests>=2.31.0

# データ処理
pandas>=2.1.0
numpy>=1.24.0

# 設定ファイル管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 進捗表示
tqdm>=4.66.0

# ロギング
colorlog>=6.7.0

# テスト
pytest>=7.4.0
pytest-cov>=4.1.0

# コード品質
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# データ分析支援（オプション）
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# テキスト類似度分析
sentence-transformers>=2.2.0

# 感情分析（日本語対応）
transformers>=4.30.0
torch>=2.0.0
fugashi>=1.3.0
unidic-lite>=1.0.8

# 日本語テキスト処理
mecab-python3>=1.0.6
janome>=0.5.0
